Fixed union sample (union now accepts Assembly names as input, not the Assemblies themselves).

+++ samples/union.rb
@@ -1,35 +1,39 @@
-require "cascading"
+#! /usr/bin/env jruby
+$:.unshift File.join(File.dirname(__FILE__), '..', 'lib')
+
+require 'cascading'
+require 'samples/cascading'
 
 input = 'output/fetched/to_be_branched.txt'
-dataUrl = "http://www.census.gov/genealogy/names/dist.all.last"
-unless File.exist? input
-  system "curl --create-dirs -o #{input} #{dataUrl}"
-end
+dataUrl = 'http://www.census.gov/genealogy/names/dist.all.last'
+system "curl --create-dirs -o #{input} #{dataUrl}" unless File.exists?(input)
 
-output = "output/union"
+output = 'output/union'
 
-flow = Cascading::Flow.new("copy_to_mysql") do
-  source "extract", tap(input)
-  sink tap(output, :replace=>true)
-    
-  assembly "extract" do
-    
-    split "line", :pattern => /[.,]*\s+/, :into=>["name", "score1", "score2", "id"], :output => ["name", "score1", "score2", "id"]
-  
-    b1 = branch "branch1" do
-      group_by "score1","name"
-      count      
-      rename ["score1"], ["score"]
+Cascading::Flow.new('copy_to_mysql') do
+  source 'extract', tap(input)
+
+  assembly 'extract' do
+    split 'line', ['name', 'score1', 'score2', 'id']
+
+    branch 'branch1' do
+      group_by 'score1', 'name' do
+        count
+      end
+      rename 'score1' => 'score'
     end
-    
-     b2 = branch "branch2" do
-      group_by "score2","name"
-      count      
-      rename ["score2"],  ["score"]      
+
+    branch 'branch2' do
+      group_by 'score2', 'name' do
+        count
+      end
+      rename 'score2' => 'score'
     end
-        
-    union b1, b2
   end
-end 
 
-flow.complete
+  assembly 'union' do
+    union 'branch1', 'branch2'
+  end
+
+  sink 'union', tap(output, :sink_mode => :replace)
+end.complete(sample_properties)
